requirements.txt
#numpy==
numpy 
pandas
python-dotenv
mysql-connector-python
pymysql
scikit-learn
sqlalchemy
#-e .  link krdega setup se